properly close process engine on application context destroy
I spent today a lot of time to debug why mockito mocks don't work with my app. And the culprit is not properly closed process engine on application context destroy.

When you run multiple test classes with different @MockBean-s Micronaut creates different ApplicationContext for each test class before test run, and recreates all beans accordingly. This is also true for ProcessEngine - this bean is recreated for each test class. However, because of not invoking ProcessEngine.close(), the scheduler thread got stuck with the first ProcessEngine which was created during the first test, and therefore it had the binding to outdated ApplicationContext with different mocks configuration for the other tests.

It seems that if you create two ProcessEngine-s in your app, without closing the first one, the scheduler thread is silently not recreated and stays bound to the first created ProcessEngine.

After this fix now my mocks config works properly.