Close #91: Sending SIGINT (=CTRL-C) to shutdown the application gracefully no longer results in an exception for in-memory h2 database.
See http://www.h2database.com/html/features.html

Quote:
"By default, a database is closed when the last connection is closed. However, if it is never closed, the database is closed when the virtual machine exits normally, using a shutdown hook. In some situations, the database should not be closed in this case, for example because the database is still used at virtual machine shutdown (to store the shutdown process in the database for example). For those cases, the automatic closing of the database can be disabled in the database URL."