Update to Micronaut 3.5.2
Note 1: Change in MnTransactionContext replaces a deprecated method which would otherwise lead to failing tests: it would fail with "No value for key [HikariDataSource (HikariPool-1)] bound to thread [Test worker]" because the rollback of the test would trigger a TransactionSynchronizationManager::unbindResource which doesn't have a corresponding TransactionSynchronizationManager::bindResource

Note 2: MnProcessEngineConfigurationTelemetryTest.kt needs to be adapted because with update to 3.5.2 the mocked bean would appear as "CamundaVersion$MockitoMock$mYxVr3eX" instead of "CamundaVersion" in the map singletonByArgumentAndQualifier in SingletonScope::findCachedSingletonBeanRegistration. By making it a real object instead of a mock we bypass that issue.